+++ src/Files.App/Actions/Favorites/PinItemAction.cs
@@ -1,27 +1,31 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
+using Files.App.Services;
+using Files.App.UserControls.Widgets;
+using Windows.Storage;
+
 namespace Files.App.Actions
 {
-	internal sealed class UnpinFolderToSidebarAction : ObservableObject, IAction
+	internal class PinItemAction : ObservableObject, IAction
 	{
 		private readonly IContentPageContext context;
 
 		private readonly IQuickAccessService service;
 
 		public string Label
-			=> "UnpinFromFavorites".GetLocalizedResource();
+			=> "PinToFavorites".GetLocalizedResource();
 
 		public string Description
-			=> "UnpinFolderToSidebarDescription".GetLocalizedResource();
+			=> "PinItemToFavoritesDescription".GetLocalizedResource();
 
 		public RichGlyph Glyph
-			=> new(opacityStyle: "ColorIconUnpinFromFavorites");
+			=> new(opacityStyle: "ColorIconPinToFavorites");
 
 		public bool IsExecutable
 			=> GetIsExecutable();
 
-		public UnpinFolderToSidebarAction()
+		public PinItemAction()
 		{
 			context = Ioc.Default.GetRequiredService<IContentPageContext>();
 			service = Ioc.Default.GetRequiredService<IQuickAccessService>();
@@ -35,11 +39,12 @@ namespace Files.App.Actions
 			if (context.HasSelection)
 			{
 				var items = context.SelectedItems.Select(x => x.ItemPath).ToArray();
-				await service.UnpinFromSidebarAsync(items);
+
+				await service.PinToSidebarAsync(items);
 			}
 			else if (context.Folder is not null)
 			{
-				await service.UnpinFromSidebarAsync(context.Folder.ItemPath);
+				await service.PinToSidebarAsync(context.Folder.ItemPath);
 			}
 		}
 
@@ -48,12 +53,14 @@ namespace Files.App.Actions
 			string[] favorites = App.QuickAccessManager.Model.FavoriteItems.ToArray();
 
 			return context.HasSelection
-				? context.SelectedItems.All(IsPinned)
-				: context.Folder is not null && IsPinned(context.Folder);
+				? context.SelectedItems.All(IsPinnable)
+				: context.Folder is not null && IsPinnable(context.Folder);
 
-			bool IsPinned(ListedItem item)
+			bool IsPinnable(ListedItem item)
 			{
-				return favorites.Contains(item.ItemPath);
+				return
+					item.PrimaryItemAttribute is StorageItemTypes.Folder &&
+					!favorites.Contains(item.ItemPath);
 			}
 		}
