+++ src/Files.App/Actions/Favorites/UnpinItemAction.cs
@@ -1,28 +1,30 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Windows.Storage;
+using Files.App.Services;
+using Files.App.UserControls.Widgets;
 
 namespace Files.App.Actions
 {
-	internal sealed class PinFolderToSidebarAction : ObservableObject, IAction
+	internal class UnpinItemAction : ObservableObject, IAction
 	{
 		private readonly IContentPageContext context;
+
 		private readonly IQuickAccessService service;
 
 		public string Label
-			=> "PinToFavorites".GetLocalizedResource();
+			=> "UnpinFromFavorites".GetLocalizedResource();
 
 		public string Description
-			=> "PinItemToFavoritesDescription".GetLocalizedResource();
+			=> "UnpinItemFromFavoritesDescription".GetLocalizedResource();
 
 		public RichGlyph Glyph
-			=> new(opacityStyle: "ColorIconPinToFavorites");
+			=> new(opacityStyle: "ColorIconUnpinFromFavorites");
 
 		public bool IsExecutable
 			=> GetIsExecutable();
 
-		public PinFolderToSidebarAction()
+		public UnpinItemAction()
 		{
 			context = Ioc.Default.GetRequiredService<IContentPageContext>();
 			service = Ioc.Default.GetRequiredService<IQuickAccessService>();
@@ -36,12 +38,11 @@ namespace Files.App.Actions
 			if (context.HasSelection)
 			{
 				var items = context.SelectedItems.Select(x => x.ItemPath).ToArray();
-
-				await service.PinToSidebarAsync(items);
+				await service.UnpinFromSidebarAsync(items);
 			}
 			else if (context.Folder is not null)
 			{
-				await service.PinToSidebarAsync(context.Folder.ItemPath);
+				await service.UnpinFromSidebarAsync(context.Folder.ItemPath);
 			}
 		}
 
@@ -50,14 +51,12 @@ namespace Files.App.Actions
 			string[] favorites = App.QuickAccessManager.Model.FavoriteItems.ToArray();
 
 			return context.HasSelection
-				? context.SelectedItems.All(IsPinnable)
-				: context.Folder is not null && IsPinnable(context.Folder);
+				? context.SelectedItems.All(IsPinned)
+				: context.Folder is not null && IsPinned(context.Folder);
 
-			bool IsPinnable(ListedItem item)
+			bool IsPinned(ListedItem item)
 			{
-				return
-					item.PrimaryItemAttribute is StorageItemTypes.Folder &&
-					!favorites.Contains(item.ItemPath);
+				return favorites.Contains(item.ItemPath);
 			}
 		}
